third_party/blink/web_tests/wpt_internal/shared_storage_selecturl_limit/run-url-selection-operation-limit.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>
<script src="../shared_storage/resources/utils.js"></script>
<script src="../fenced_frame/resources/utils.js"></script>

<body>
<script>
'use strict';

promise_test(async () => {
  const ancestorKey = token();
  const urls = generateUrls(8, "../shared_storage/resources/frame",
                            [ancestorKey]);

  await sharedStorage.worklet.addModule("../shared_storage/resources/simple-module.js");

  for (let i = 0; i < 8; ++i) {
    let uuid = await sharedStorage.selectURL(
        "test-url-selection-operation", urls,
        {data: {'mockResult': 1}, keepAlive: true});
    assert_true(uuid.startsWith('urn:uuid:'));
    attachFencedFrame(uuid, 'opaque-ads');
    const result = await nextValueFromServer(ancestorKey);
    assert_equals(result, "frame1_loaded");
  }

  // The per-origin per-pageload bit limit for `selectURL()` has been
  // reached. The next call should return the default URL.
  let uuid = await sharedStorage.selectURL(
      "test-url-selection-operation", urls,
      {data: {'mockResult': 1}});
  assert_true(uuid.startsWith('urn:uuid:'));
  attachFencedFrame(uuid, 'opaque-ads');
  const result = await nextValueFromServer(ancestorKey);
  assert_equals(result, "frame0_loaded");

}, 'selectURL() with per-origin per-pageload limit');
</script>
</body>
